+++ SeedShop/src/main/java/com/apu/seedshop/repository/UserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.apu.seedshop.jpa.Appuser;
 import org.springframework.data.repository.query.Param;
 
-public interface UserRepository extends CrudRepository<Appuser, Integer> {
+public interface UserRepository extends CrudRepository<Appuser, Long> {
     public List<Appuser> findByFirstName(@Param("firstName") String firstName);
     public List<Appuser> findBySecNameAndFirstName(String secName, String firstName);
     @Override
